Review (#388)
* Bug is fixed and added new code

* new code for the table

* Edits in markdown file

* some edits in test

* Bugs fix

* Codecov

* I cleaned up my code and separated classes to make it easier to work with. It is not ready yet; if Codecov fails, I will include more tests.

* forgot black

* flake8

* bug fix

* Edits codes according to the comments

* Edited codes according to the comments in the GitHub

* Defined new function in stability_simulation.py to check stability for
given points and excluded codecov function that generates a table.

* small edits for codecov

* removed no cover
