+++ pySDC/projects/Second_orderSDC/check_data_folder.py
@@ -3,7 +3,7 @@ import os
 folder_name = "./data"
 
 # Check if the folder already exists
-if not os.path.isdir(folder_name):  # pragma: no cover
+if not os.path.isdir(folder_name):
     # Create the folder
     os.makedirs(folder_name)
 else:
